Fixed all issues

+++ db/migrate/20210303184847_add_avarage_rating_to_products.rb
@@ -1,5 +1,5 @@
 class AddAvarageRatingToProducts < ActiveRecord::Migration[6.1]
   def change
-    add_column :products, :average_rating, :double
+    add_column :products, :average_rating, :numeric
   end
 end
